Enum kecc::asm::RType

pub enum RType {

    Add(DataSize),
    Sub(DataSize),
    Sll(DataSize),
    Srl(DataSize),
    Sra(DataSize),
    Mul(DataSize),
    Div {
        data_size: DataSize,
        is_signed: bool,
    },
    Rem {
        data_size: DataSize,
        is_signed: bool,
    },
    Slt {
        is_signed: bool,
    },
    Xor,
    Or,
    And,
    Fadd(DataSize),
    Fsub(DataSize),
    Fmul(DataSize),
    Fdiv(DataSize),
    Feq(DataSize),
    Flt(DataSize),
    FmvIntToFloat {
        float_data_size: DataSize,
    },
    FmvFloatToInt {
        float_data_size: DataSize,
    },
    FcvtIntToFloat {
        int_data_size: DataSize,
        float_data_size: DataSize,
        is_signed: bool,
    },
    FcvtFloatToInt {
        float_data_size: DataSize,
        int_data_size: DataSize,
        is_signed: bool,
    },
    FcvtFloatToFloat {
        from: DataSize,
        to: DataSize,
    },

}

If the enum variant contains bool, It means that different instructions exist depending on whether the operand is signed or not.

Variants

Add(DataSize)

Sub(DataSize)

Sll(DataSize)

Srl(DataSize)

Sra(DataSize)

Mul(DataSize)

Div

Fields

data_size: DataSize

is_signed: bool

Rem

Fields

data_size: DataSize

is_signed: bool

Slt

Fields

is_signed: bool

Xor

Or

And

Fadd(DataSize)

Fsub(DataSize)

Fmul(DataSize)

Fdiv(DataSize)

Feq(DataSize)

Flt(DataSize)

FmvIntToFloat

Fields

float_data_size: DataSize

fmv.w.x or fmv.d.x

FmvFloatToInt

Fields

float_data_size: DataSize

fmv.x.w or fmv.x.w

FcvtIntToFloat

Fields

int_data_size: DataSize

float_data_size: DataSize

is_signed: bool

fcvt.s.l(u) or fcvt.d.l(u) fcvt.s.w(u) or fcvt.d.w(u)

FcvtFloatToInt

Fields

float_data_size: DataSize

int_data_size: DataSize

is_signed: bool

fcvt.l(u).s or fcvt.l(u).d fcvt.w(u).s or fcvt.w(u).d

FcvtFloatToFloat

Fields

from: DataSize

to: DataSize

fcvt.s.d or fcvt.d.s

Implementations

impl RType

pub fn add(dtype: Dtype) -> Self

pub fn sub(dtype: Dtype) -> Self

pub fn sll(dtype: Dtype) -> Self

pub fn srl(dtype: Dtype) -> Self

pub fn sra(dtype: Dtype) -> Self

pub fn mul(dtype: Dtype) -> Self

pub fn div(dtype: Dtype, is_signed: bool) -> Self

pub fn rem(dtype: Dtype, is_signed: bool) -> Self

pub fn fadd(dtype: Dtype) -> Self

pub fn fsub(dtype: Dtype) -> Self

pub fn fmul(dtype: Dtype) -> Self

pub fn fdiv(dtype: Dtype) -> Self

pub fn feq(dtype: Dtype) -> Self

pub fn flt(dtype: Dtype) -> Self

pub fn fmv_int_to_float(dtype: Dtype) -> Self

pub fn fmv_float_to_int(dtype: Dtype) -> Self

pub fn fcvt_int_to_float(from: Dtype, to: Dtype) -> Self

pub fn fcvt_float_to_int(from: Dtype, to: Dtype) -> Self