Struct kecc::ir::dtype::BaseDtype

struct BaseDtype {
    scalar: Option<TypeSpecifier>,
    size_modifiers: Vec<TypeSpecifier>,
    signed_option: Option<TypeSpecifier>,
    typedef_name: Option<String>,
    struct_type: Option<StructType>,
    is_const: bool,
    is_typedef: bool,
}

Fields
scalar: Option<TypeSpecifier>size_modifiers: Vec<TypeSpecifier>signed_option: Option<TypeSpecifier>typedef_name: Option<String>struct_type: Option<StructType>is_const: boolis_typedef: bool
Implementations

impl BaseDtype

fn apply_storage_class(
&mut self,
storage_class: &StorageClassSpecifier
) -> Result<(), DtypeError>

Apply StorageClassSpecifier to BaseDtype.

Let’s say declaration is typedef int i32_t;, if self represents int and type_qualifier represents typedef, self is transformed to representing typedef int.

Arguments

self - Part that has been converted to ‘BaseDtype’ on the declaration.

storage_class - storage class requiring to apply to ‘self’ immediately.

fn apply_type_specifier(
&mut self,
type_specifier: &TypeSpecifier
) -> Result<(), DtypeError>

Apply TypeSpecifier to BaseDtype.

Let’s say the declaration is const int a;, if self represents int and type_specifier represents const, self is transformed to representing const int.

Arguments

self - Part that has been converted to ‘BaseDtype’ on the declaration.

type_qualifier - type qualifiers requiring to apply to ‘self’ immediately.

fn apply_type_qualifier(
&mut self,
type_qualifier: &TypeQualifier
) -> Result<(), DtypeError>

Apply Typequalifier to BaseDtype.

Let’s say the declaration is const int a;, if self represents int and type_qualifier represents const, self is transformed to representing const int.

Arguments

self - Part that has been converted to ‘BaseDtype’ on the declaration.

type_qualifier - type qualifiers requiring to apply to ‘self’ immediately.

pub fn apply_specifier_qualifier(
&mut self,
typename_specifier: &SpecifierQualifier
) -> Result<(), DtypeError>

pub fn apply_declaration_specifier(
&mut self,
declaration_specifier: &DeclarationSpecifier
) -> Result<(), DtypeError>

pub fn apply_pointer_qualifier(
&mut self,
pointer_qualifier: &PointerQualifier
) -> Result<(), DtypeError>

Apply PointerQualifier to BaseDtype.

let’s say pointer declarator is * const of const int * const a;. If self represents nothing, and pointer_qualifier represents const between the first and second asterisk, self is transformed to representing const. This information is used later when generating Dtype.

Arguments

self - Part that has been converted to ‘BaseDtype’ on the pointer declarator.

pointer_qualifier - Pointer qualifiers required to apply to ‘BaseDtype’ immediately.

pub fn apply_specifier_qualifiers(
&mut self,
typename_specifiers: &[Node<SpecifierQualifier>]
) -> Result<(), DtypeError>

pub fn apply_declaration_specifiers(
&mut self,
declaration_specifiers: &[Node<DeclarationSpecifier>]
) -> Result<(), DtypeError>