kecc::ir
Enum Dtype

pub enum Dtype {
    Unit {
        is_const: bool,
    },
    Int {
        width: usize,
        is_signed: bool,
        is_const: bool,
    },
    Float {
        width: usize,
        is_const: bool,
    },
    Pointer {
        inner: Box<Dtype>,
        is_const: bool,
    },
    Array {
        inner: Box<Dtype>,
        size: usize,
    },
    Struct {
        name: Option<String>,
        fields: Option<Vec<Named<Dtype>>>,
        is_const: bool,
        size_align_offsets: Option<(usize, usize, Vec<usize>)>,
    },
    Function {
        ret: Box<Dtype>,
        params: Vec<Dtype>,
    },
    Typedef {
        name: String,
        is_const: bool,
    },
}

TODO(document)

Variants

Unit

TODO(document)

Fields

is_const: bool

TODO(document)

Int

TODO(document)

Fields

width: usize

TODO(document)

is_signed: bool

TODO(document)

is_const: bool

TODO(document)

Float

TODO(document)

Fields

width: usize

TODO(document)

is_const: bool

TODO(document)

Pointer

TODO(document)

Fields

inner: Box<Dtype>

TODO(document)

is_const: bool

TODO(document)

Array

TODO(document)

Fields

inner: Box<Dtype>

TODO(document)

size: usize

TODO(document)

Struct

TODO(document)

Fields

name: Option<String>

TODO(document)

fields: Option<Vec<Named<Dtype>>>

TODO(document)

is_const: bool

TODO(document)

size_align_offsets: Option<(usize, usize, Vec<usize>)>

TODO(document)

Function

TODO(document)

Fields

ret: Box<Dtype>

TODO(document)

params: Vec<Dtype>

TODO(document)

Typedef

TODO(document)

Fields

name: String

TODO(document)

is_const: bool

TODO(document)

Implementations

impl Dtype

pub const BITS_OF_BYTE: usize = 8usize

TODO(document)

pub const SIZE_OF_BYTE: usize = 1usize

TODO(document)

pub const SIZE_OF_POINTER: usize = 8usize

TODO(document)

pub const SIZE_OF_CHAR: usize = 1usize

TODO(document)

pub const SIZE_OF_SHORT: usize = 2usize

TODO(document)

pub const SIZE_OF_INT: usize = 4usize

TODO(document)

pub const SIZE_OF_LONG: usize = 8usize

TODO(document)

pub const SIZE_OF_LONGLONG: usize = 8usize

TODO(document)

pub const SIZE_OF_FLOAT: usize = 4usize

TODO(document)

pub const SIZE_OF_DOUBLE: usize = 8usize

TODO(document)

pub const BOOL: Self

TODO(document) A boolean value cannot be signed.

pub const CHAR: Self

TODO(document)

pub const SHORT: Self

TODO(document)

pub const INT: Self

TODO(document)

pub const LONG: Self

TODO(document)

pub const LONGLONG: Self

TODO(document)

pub const FLOAT: Self

TODO(document)

pub const DOUBLE: Self

TODO(document)

pub const fn unit() -> Self

TODO(document)

pub const fn int(width: usize) -> Self

TODO(document)

pub const fn float(width: usize) -> Self

TODO(document)

pub fn pointer(inner: Dtype) -> Self

TODO(document)

pub fn array(base_dtype: Dtype, size: usize) -> Self

TODO(document)

Examples

Suppose the C declaration is int *a[2][3]. Then a’s ir::Dtype should be [2 x [3 x int*]]. But in the AST, it is parsed as Array(3, Array(2, Pointer(int))), reversing the order of 2 and 3. In the recursive translation of a declaration into Dtype, we need to insert 3 inside [2 * int*].

pub fn structure(name: Option<String>, fields: Option<Vec<Named<Self>>>) -> Self

TODO(document)

pub fn fill_size_align_offsets_of_struct( self, structs: &HashMap<String, Option<Dtype>>, ) -> Result<Self, DtypeError>

pub fn function(ret: Dtype, params: Vec<Dtype>) -> Self

pub fn typedef(name: String) -> Self

pub fn get_int_width(&self) -> Option<usize>

pub fn get_float_width(&self) -> Option<usize>

pub fn get_pointer_inner(&self) -> Option<&Self>

pub fn get_array_inner(&self) -> Option<&Self>

pub fn get_struct_name(&self) -> Option<&Option<String>>

pub fn get_struct_fields(&self) -> Option<&Option<Vec<Named<Self>>>>

pub fn get_struct_size_align_offsets( &self, ) -> Option<&Option<(usize, usize, Vec<usize>)>>

pub fn get_function_inner(&self) -> Option<(&Self, &Vec<Self>)>

pub fn is_scalar(&self) -> bool

pub fn is_int_signed(&self) -> bool

pub fn is_const(&self) -> bool

pub fn is_immutable(&self, structs: &HashMap<String, Option<Dtype>>) -> bool

Check if Dtype is constant. if it is constant, the variable of Dtype is not assignable.

fn is_immutable_for_array_struct_field_inner( &self, structs: &HashMap<String, Option<Dtype>>, ) -> bool

pub fn set_const(self, is_const: bool) -> Self

pub fn size_align_of( &self, structs: &HashMap<String, Option<Dtype>>, ) -> Result<(usize, usize), DtypeError>

pub fn get_offset_struct_field( &self, field_name: &str, structs: &HashMap<String, Option<Dtype>>, ) -> Option<(usize, Self)>

pub fn set_signed(&self, is_signed: bool) -> Self

pub fn try_from_ast_declaration_specifiers( specifiers: &[Node<DeclarationSpecifier>], ) -> Result<(Self, bool), DtypeError>

Derive a data type from declaration specifiers.

pub fn try_from_ast_struct_declaration( declaration: &StructDeclaration, ) -> Result<Vec<Named<Self>>, DtypeError>

Derive a data type and its name from the struct declaration.

pub fn with_ast_declarator( self, declarator: &Declarator, ) -> Result<Named<Self>, DtypeError>

Generate Dtype based on declarator and self which has a scalar type.

Let’s say declaration is const int * const * const a;. In general self start with const int which has a scalar type and declarator represents * const * const with ast::Declarator.

Arguments

declarator - Parts requiring conversion to ‘Dtype’ on the declaration.

pub fn with_ast_array_size( self, array_size: &ArraySize, ) -> Result<Self, DtypeError>

Generates Dtype based on declarator and self which has a scalar type.

Let’s say the AST declaration is int a[2][3]; self represents int [2]; and array_size is [3]. Then this function should return int [2][3].

Arguments

array_size - the array size to add to self.

pub fn resolve_typedefs( self, typedefs: &HashMap<String, Dtype>, ) -> Result<Self, DtypeError>

pub fn resolve_structs( self, structs: &mut HashMap<String, Option<Dtype>>, tempid_counter: &mut usize, ) -> Result<Self, DtypeError>

If the struct type has a definition, it is saved to the struct table and transformed to a struct type with no definition.