Struct kecc::Irgen

pub struct Irgen {
    decls: BTreeMap<String, Declaration>,
    typedefs: HashMap<String, Dtype>,
    structs: HashMap<String, Option<Dtype>>,
    struct_tempid_counter: usize,
}

Fields
decls: BTreeMap<String, Declaration>typedefs: HashMap<String, Dtype>structs: HashMap<String, Option<Dtype>>struct_tempid_counter: usize
Implementations

impl Irgen

const BID_INIT: BlockId = _

const BID_COUNTER_INIT: usize = 1usize

const TEMPID_COUNTER_INIT: usize = 0usize

fn add_declaration(&mut self, source: &Declaration) -> Result<(), IrgenError>

Add a declaration. It can be either a struct, typedef, or a variable.

fn add_function_definition(
&mut self,
source: &FunctionDefinition
) -> Result<(), IrgenError>

Add a function definition.

fn add_decl(&mut self, name: &str, decl: Declaration) -> Result<(), IrgenError>

Adds a possibly existing declaration.

Returns error if the previous declearation is incompatible with decl.