kecc::write_base
Trait WriteLine

pub trait WriteLine {
    // Required method
    fn write_line(&self, indent: usize, write: &mut dyn Write) -> Result<()>;
}

A trait for writing a type to a Write stream with a new line.

Required Methods

fn write_line(&self, indent: usize, write: &mut dyn Write) -> Result<()>

Write self to write, starting at indent number of double spaces, with a newline at the ned.