kecc::write_base
Trait WriteString

pub trait WriteString {
    // Required method
    fn write_string(&self) -> String;
}

Format types to a String.

Most cases, fmt::Display is used to format a type to a string. However, in some cases, we can’t implement fmt::Display for a type as it is defined in another crate. In such cases, we can implement this trait to format the type to a string.

Required Methods

fn write_string(&self) -> String

Change a type into a String.