Crate kecc

KECC: KAIST Educational C Compiler.

Re-exports

pub use ir::Parse as IrParse;

pub use ir::Visualizer as IrVisualizer;

Modules

asm

asmgen 🔒

c 🔒

ir

The intermediate representation.

irgen 🔒

Homework: IR Generation

opt 🔒

tests 🔒

utils 🔒

write_base 🔒

Macros

ok_or

Ok or executing the given expression.

ok_or_exit

Ok or exiting the process.

some_or

Some or executing the given expression.

some_or_exit

Ok or exiting the process.

Structs

Asmgen

FunctionPass

Irgen

Parse

TODO(document)

Repeat

SimplifyCfgConstProp

Simplifies block exits by propagating constants.

SimplifyCfgEmpty

Removes empty blocks

SimplifyCfgMerge

Merges two blocks if a block is pointed to only by another

SimplifyCfgReach

Retains only those blocks that are reachable from the init.

Traits

AssertSupported

TODO(document)

IsEquiv

TODO(document)

Optimize

Translate

TODO(document)

Functions

test_asmgen

Tests asmgen.

test_end_to_end

Tests end-to-end translation.

test_irgen

Tests irgen.

test_irparse

Tests irparse.

test_opt

Tests optimizations.

test_write_c

Tests write_c.

write

TODO(document)

Type Aliases

Deadcode

Gvn

Mem2reg

O0

O1

SimplifyCfg