kecc
Module asm

Modules

write_asm 🔒

Structs

Asm

An assembly file.

Block

Function

Label

Label is used as branch, unconditional jump targets and symbol offsets.

Section

TranslationUnit

Variable

Enums

BType

DataSize

Directive

The assembler implements several directives that control the assembly of instructions into an object file.

IType

Immediate

Instruction

RISC-V Base Instructions Set.

Pseudo

The assembler implements several convenience psuedo-instructions that are formed from multiple instructions in the base ISA, but have implicit arguments or reversed arguments that result in distinct semantics.

RType

If the enum variant contains is_signed : bool, it corresponds to different instructions depending on whether the operand is signed or not.

Register

ABI name for RISC-V integer and floating-point register.

RegisterType

RelocationFunction

The relocation function creates synthesize operand values that are resolved at program link time and are used as immediate parameters for specific instructions.

SType

SectionType

SymbolType

UType